+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,21 +2,26 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.FilmorateValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class FilmService {
 
+    private static final LocalDate cinemaBirthday = LocalDate.of(1895, 12, 28);
     private final FilmStorage storage;
     private final UserService userService;
 
     public Film create(Film film) {
+        validateReleaseDate(film);
         return storage.create(film);
     }
+
 
     public List<Film> getAllFilms() {
         return storage.getAllFilms();
@@ -27,6 +32,7 @@ public class FilmService {
     }
 
     public Film update(Film film) {
+        validateReleaseDate(film);
         return storage.update(film);
     }
 
@@ -44,5 +50,13 @@ public class FilmService {
 
     public List<Film> getPopularFilms(int count) {
         return storage.getPopular(count);
+    }
+
+
+    private void validateReleaseDate(Film film) {
+        if (film.getReleaseDate().isBefore(cinemaBirthday)) {
+            throw new FilmorateValidationException(
+                    String.format("Дата релиза не может быть раньше чем %s", cinemaBirthday));
+        }
     }
 }
